Add several cells in dataTypesPython and data directory
Several cells were added in the dataTypesPython jupyter notebook. The
first part explain syntaxes of python language, functions and strings
The second part show the use of lists, tupples and dictionaries
Finally It was added a example of csv files manipulation whit some
operations for exploration data.

In order to manipulated data in csv files. A new directory was added
for save data in deferents formats, the first file is mpg.csv witch has
data of cars.