Add numpy jupyter notebook, winequality-red.csv dataset and image
dataTypesPython.ipynb has new cells for datetime, time, map, objects,
etc. objects in python.
numpy.ipynb is new jupyter notebook which has instructions for use
numpy library, arrays and operations.
winequality-red-csv is a dataset with data for wines an their quality
this is used in numpy.ipynb
chis.tiff is an image which is used in numpy.ipynb to manipulate image
as example to data abstraction.